Check for `test.only` before push
This check tests if a developer added `test.only` to a spec before
pushing.

This commit also fixes tests that were failing but didn't error because
of the mistaken test.only call.

In message sorting, the logic is that

1) Deletions should precede additions

2) Deletions of files or directories that are within other directories
should proceed deletions of the outer directories. (& vice versa for
additions)

3) For deletions of files that are in directories that are inside the
directories of other files, the inner-most files are deleted before
the other-most, similar to the deletion of nested directories (and vice
versa for additions)

eg `A/B/topic1.expl` should be deleted after `A/B/C/topic2.expl` because
`A/B` contains `A/B/C`

4) If we are not dealing with folders that are in other folders, or
files that are in directories that are within another file's
directory, ie neither path has a "containment" relationship with the
other, then we just alphabetize, for additions and deletions.

+++ cli/commands/bulk/file_system_change_calculator.js
@@ -65,7 +65,9 @@ class FileSystemChangeCalculator {
 
     this.originalSelectionFileSet.directories.forEach(originallySelectedDirectory => {
       if (!this.newFileSet.hasDirectory(originallySelectedDirectory.path)) { // a directory was selected that got deleted
-        fileSystemChange.deleteDirectory(originallySelectedDirectory.path);
+        if (!this.diskFileSet.files.find(f => f.path.startsWith(originallySelectedDirectory.path) && !f.path.endsWith('.expl'))) { // non-expl files in directory prevent directory deletion
+          fileSystemChange.deleteDirectory(originallySelectedDirectory.path);
+        }
       }
     });
